==> 02_preprocessing.ipynb <==
## Pre-processing to split the dataset into train and test dataset

# Create data/input directory if it is not exists in current directory
!mkdir -p data/input

# Create data/preprocess director if it is not exists in current directory
!mkdir -p data/preprocess

# Download input dataset into local directory
!wget https://storage.googleapis.com/advanced-solutions-lab/fraud/creditcard.csv -O ./data/input/creditcard.csv

# import library functions
import pandas as pd

# Pandas display column upto 50
pd.set_option("display.max_columns",50)

df = pd.read_csv('./data/input/creditcard.csv')

df.head()

df.describe()

#Drop all of the features that have very similar distributions between the two types of transactions.
df = df.drop(['V28','V27','V26','V25','V24','V23','V22','V20','V15','V13','V8'], axis =1)

df.head()

#Create dataframes of only Fraud and Normal transactions.
df_fraud = df[df.Class == 1]
df_normal = df[df.Class == 0]

df_fraud.head()

df_normal.head()

# Set df_train equal to 80% of the fraudulent transactions.
df_train = df_fraud.sample(frac=0.8)

# Add 80% of the normal transactions to df_train.
df_train = pd.concat([df_train, df_normal.sample(frac = 0.8)], axis = 0)

# df_test contains all the transaction not in df_train.
df_test = df.loc[~df.index.isin(df_train.index)]

df_train.head()

df_test.head()

def transform_feature_cols(df):
  """
  Transform each feature in features so that it has a mean of 0 and standard deviation of 1; 
  This helps with training the neural network.
  """
  features = df.columns.values
  for feature in features:
    if feature != 'Class':
      mean, std = df[feature].mean(), df[feature].std()
      df.loc[:, feature] = (df[feature] - mean) / std
  return df

df_train.loc[:] = transform_feature_cols(df_train)

df_test.loc[:] = transform_feature_cols(df_test)

<b>Preprocessed files with header<b>

df_train.to_csv('./data/preprocess/creditcard_header_train.csv', sep=',', encoding='utf-8', index=False)
df_test.to_csv('./data/preprocess/creditcard_header_test.csv', sep=',', encoding='utf-8', index=False)

<b>Preprocessed files without header - required for google cloud ml and read_dataset with csv file format</b>

df_train.to_csv('./data/preprocess/creditcard_train.csv', sep=',', encoding='utf-8', header=False, index=False)
df_test.to_csv('./data/preprocess/creditcard_test.csv', sep=',', encoding='utf-8', header=False, index=False)

df_train = pd.read_csv('./data/preprocess/creditcard_header_train.csv')
df_train.head()

df_train.Class.unique()

df_train['Class'].value_counts()

df_test = pd.read_csv('./data/preprocess/creditcard_header_test.csv')
df_test.head()

df_test.Class.unique()

df_test['Class'].value_counts()

<b>Preprocessing for having one hot encoding for class label <b>

df_train.loc[df_train.Class == 0, 'Normal'] = 1
df_train.loc[df_train.Class == 1, 'Normal'] = 0
df_train_ohe = df_train.rename(columns={'Class': 'Fraud'})
df_train_ohe.head()

df_test.loc[df_test.Class == 0, 'Normal'] = 1
df_test.loc[df_test.Class == 1, 'Normal'] = 0
df_test_ohe = df_test.rename(columns={'Class': 'Fraud'})
df_test_ohe.head()

df_train_ohe=df_train_ohe.astype({"Normal": int, "Fraud": int})
df_train_ohe.head()

df_train_ohe.Fraud.unique()

df_train_ohe['Fraud'].value_counts()

df_train_ohe.Normal.unique()

df_train_ohe['Normal'].value_counts()

df_test_ohe=df_test_ohe.astype({"Normal": int, "Fraud": int})
df_test_ohe.head()

df_test_ohe.Fraud.unique()

df_test_ohe['Fraud'].value_counts()

df_test_ohe.Normal.unique()

df_test_ohe['Normal'].value_counts()

<b>Write to CSV files<b>

df_train_ohe.to_csv('./data/preprocess/creditcard_ohe_header_train.csv', sep=',', encoding='utf-8', index=False)
df_test_ohe.to_csv('./data/preprocess/creditcard_ohe_header_test.csv', sep=',', encoding='utf-8', index=False)

df_train_ohe.to_csv('./data/preprocess/creditcard_ohe_train.csv', sep=',', encoding='utf-8', header=False, index=False)
df_test_ohe.to_csv('./data/preprocess/creditcard_ohe_test.csv', sep=',', encoding='utf-8', header=False, index=False)

<pre>
# Copyright 2018 Atos. All Rights Reserved.
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#      http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.
</pre>
